+++ cli/ava.config.js
@@ -1,0 +1,6 @@
+export default {
+  extensions: {
+    ts: 'module',
+  },
+  files: ['**/__tests__/**/*.spec.ts'],
+}