+++ cli/src/commands/new.ts
@@ -94,10 +94,6 @@ export class NewCommand extends BaseNewCommand {
   }
 
   private async fetchTargets(): Promise<TargetTriple[]> {
-    if (this.enableDefaultTargets) {
-      return DEFAULT_TARGETS.concat()
-    }
-
     if (this.enableAllTargets) {
       return AVAILABLE_TARGETS.concat()
     }
@@ -107,7 +103,7 @@ export class NewCommand extends BaseNewCommand {
       type: 'checkbox',
       loop: false,
       message: 'Choose target(s) your crate will be compiled to',
-      default: DEFAULT_TARGETS,
+      default: this.enableDefaultTargets ? DEFAULT_TARGETS : [],
       choices: AVAILABLE_TARGETS,
     })