+++ examples/napi/src/date.rs
@@ -1,3 +1,5 @@
+use std::str::FromStr;
+
 use chrono::{Duration, Utc};
 use napi::bindgen_prelude::*;
 
@@ -20,4 +22,14 @@ fn chrono_date_add_1_minute(input: chrono::DateTime<Utc>) -> chrono::DateTime<Ut
 pub struct Dates {
   pub start: chrono::DateTime<Utc>,
   pub end: Option<chrono::DateTime<Utc>>,
+}
+
+#[napi]
+pub fn chrono_native_date_time(date: chrono::NaiveDateTime) -> i64 {
+  date.timestamp_millis()
+}
+
+#[napi]
+pub fn chrono_native_date_time_return() -> Option<chrono::NaiveDateTime> {
+  chrono::NaiveDateTime::from_str("2016-12-23T15:25:59.325").ok()
 }